+++ browser/deferral.js
@@ -7,16 +7,22 @@ function needs (action) {
   var demands = [];
   var is = deferred(action, state.deferrals);
   if (is) {
-    if (invalid(state.templates)) {
+    if (invalid('templates')) {
       demands.push('template');
     }
-    if (invalid(state.controllers)) {
+    if (invalid('controllers')) {
       demands.push('controller');
     }
   }
 
-  function invalid (store) {
-    return !(action in store) || store[action].version !== state.version;
+  function invalid (type) {
+    var store = state[type];
+    var invalid = !store[action] || store[action].version !== state.version;
+    if (invalid) {
+      global.DEBUG && global.DEBUG('[deferral] deferred %s %s not found', action, type.substr(0, type.length - 1));
+      return true;
+    }
+    return false;
   }
 
   return demands;